requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-filter>=24.2
djangorestframework-simplejwt>=5.3,<5.4
dj-database-url>=2.1,<2.2
django-cors-headers>=4.4,<4.5
psycopg[binary]>=3.2,<3.3
python-dotenv>=1.0,<1.1
